+++ glassure/optimization.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 
 from copy import deepcopy
+from typing import Optional
 
 import numpy as np
+from scipy.integrate import simpson
 from lmfit import Parameters, minimize
 
 from . import Pattern
-from .transform import calculate_fr, calculate_gr
+from .transform import calculate_fr, calculate_gr, calculate_sq_from_fr
 
 __all__ = [
     "optimize_sq",
@@ -19,6 +21,7 @@ def optimize_sq(
     r_cutoff: float,
     iterations: int,
     atomic_density: float,
+    r_step: float = 0.01,
     use_modification_fcn: bool = False,
     attenuation_factor: float = 1,
     fcn_callback=None,
@@ -38,6 +41,9 @@ def optimize_sq(
         number of back and forward transforms
     :param atomic_density:
         density in atoms/A^3
+    :param r_step:
+        step size for the r-axis, default is 0.01. Use smaller values for better accuracy (especially if needed for
+         fft)
     :param use_modification_fcn:
         Whether to use the Lorch modification function during the Fourier transform.
         Warning: When using the Lorch modification function, usually more iterations are needed to get to the
@@ -58,7 +64,7 @@ def optimize_sq(
     :return:
         optimized S(Q) pattern
     """
-    r = np.arange(0, r_cutoff, 0.02)
+    r = np.arange(0, r_cutoff, r_step)
     sq_pattern = deepcopy(sq_pattern)
     for iteration in range(iterations):
         fr_pattern = calculate_fr(
@@ -69,11 +75,17 @@ def optimize_sq(
 
         delta_fr = fr_int + 4 * np.pi * r * atomic_density
 
-        in_integral = np.array(np.sin(np.outer(q.T, r))) * delta_fr
-        integral = np.trapz(in_integral, r) / attenuation_factor
-        sq_optimized = sq_int * (1 - 1.0 / q * integral)
+        if fourier_transform_method == "fft":
+            sq_trans_fft = (
+                calculate_sq_from_fr(Pattern(r, delta_fr), sq_pattern.x, method="fft")
+                - 1
+            )
+            iq = sq_trans_fft.y
+        else:
+            in_integral = np.array(np.sin(np.outer(q.T, r))) * delta_fr
+            iq = np.trapz(in_integral, r) / q
 
-        sq_pattern = Pattern(q, sq_optimized)
+        sq_pattern = sq_pattern * (1 - iq / attenuation_factor)
 
         if fcn_callback is not None and iteration % callback_period == 0:
             fr_pattern = calculate_fr(
@@ -94,24 +106,27 @@ from .methods import ExtrapolationMethod
 def optimize_density(
     data_config: DataConfig,
     calculation_config: CalculationConfig,
-    type: str = "gr",
-    min_range: tuple[float, float] = (0, 1),
-    method: str = "lsq",
-) -> tuple[float, float]:
+    method: str = "fr",
+    min_range: Optional[tuple[float, float]] = None,
+    vary_bkg_scaling: bool = True,
+    bkg_limits: tuple[float, float] = (0.9, 1.1),
+    optimization_method: str = "lsq",
+) -> tuple[float, float, float, float]:
     """
-    Optimizes the density of the sample using the g(r) or S(Q) (chosen by the method parameter). The density in the
-    Sample configuration of the CalculationConfig is taking as starting parameter
+    Optimizes the density of the sample using the g(r), f(r) or S(Q) (chosen by the method parameter).
+    The density in the SampleConfig of the DataConfig is taking as starting parameter
 
-    For type='gr' the optimization is based on the g(r) function, and the density is optimized to minimize the
-    low g(r) region to be close to zero. For better results, the g(r) function is calculated with the Lorch
-    modification function. The general procedure is explained in Eggert et al. 2002 PRB, 65, 174105.
+    For method='gr' or method='fr' the optimization is based on the g(r) or f(r) function, and the density is
+    optimized to minimize the low g(r) or f(r) region to be close to zero. For better results, the g(r) function
+    is calculated with the Lorch modification function. The general procedure is explained in
+    Eggert et al. 2002 PRB, 65, 174105.
 
-    For type='sq' the optimization is based on the low Q part of the S(Q) function, and the density is optimized
+    For method='sq' the optimization is based on the low Q part of the S(Q) function, and the density is optimized
     to minimize the difference between the original S(Q) function without any optimization and the optimized S(Q)
     function. The configuration should have extrapolation enabled for this to work best.
     For polyatomic systems, finding the density using this procedure is much less susceptible to the Q_max value of
     the S(Q) than the g(r) based optimization. However, density is not exactly the same for both methods and the
-    method needs to be verified further. (Please us the type='sq' with caution.)
+    method needs to be verified further. (Please us the method='sq' with caution.)
 
     The best for both types is to have a reference density to compare it to. Based on this then further calculations
     of e.g. high pressure or high temperature densities can be performed.
@@ -130,67 +145,113 @@ def optimize_density(
     calculation_config.transform.extrapolation.method = ExtrapolationMethod.LINEAR
     calculation_config.optimize = OptimizeConfig(r_cutoff=1.4)
 
-    density, error = optimize_density(data_config, calculation_config, type='gr', range=(0.1, 1.2))
+    density, density_error, bkg_scaling, bkg_error = optimize_density(data_config, calculation_config, method='gr', range=(0.1, 1.2))
     ```
 
     :param data_config:
         Data configuration
     :param calculation_config:
         Calculation configuration
-    :param type:
-        Method to use for the optimization. Possible values are 'gr' and 'sq'.
+    :param method:
+        Method to use for the optimization. Possible values are 'gr', 'fr' and 'sq'.
     :param min_range:
-        x range of the data to use for the minimization to find the density. For method='gr' this is the r-range of the
-        g(r) function to minimize to be close to zero. For method='sq' this is the Q-range of the S(Q) function to
-        minimize the difference between the original and optimized S(Q) function.
+        x range of the data to use for the minimization to find the density. For method='gr' and 'fr this is the
+        r-range of the g(r)/f(r) function to minimize. For method='sq' this is the Q-range of the S(Q) function to
+        minimize the difference between the original and optimized S(Q) function. Default is None which means that
+        the range is (0, calculation_config.optimize.r_cutoff) for method='gr' and 'fr' and
+        (0, calculation_config.transform.q_max) for method='sq'.
+    :param vary_bkg_scaling:
+        Whether to vary the background scaling during the optimization. Default is True.
+    :param bkg_limits:
+        relative limits for the background scaling. The background scaling is optimized to be within these limits.
+        Default is (0.9, 1.1) which means that the background scaling is optimized to be within 10% of the starting
+        value.
     :param optimization_method:
         Method to use for the optimization. Possible values are 'nelder' and 'lsq'.
 
     :return:
-        a tuple with two values:
-        - the density and the standard error for optimization_method='lsq'
-        - the density and the residual for optimization_method='nelder'
+        a tuple with four values:
+        - the density, the standard error, the background scaling and the standard error for optimization_method='lsq'
     """
+
+    if calculation_config.optimize is None and min_range is None and not method == "sq":
+        raise ValueError(
+            "For optimizing density using 'gr' or 'fr' the calculation configuration needs to have the "
+            "optimize configuration or the min_range parameter needs to be set."
+        )
 
     params = Parameters()
     params.add("density", value=calculation_config.sample.density, min=0.0, max=100)
+    params.add(
+        "bkg_scaling",
+        value=data_config.bkg_scaling,
+        min=data_config.bkg_scaling * bkg_limits[0],
+        max=data_config.bkg_scaling * bkg_limits[1],
+        vary=vary_bkg_scaling,
+    )
 
     optim_config = calculation_config.model_copy(deep=True)
-    optim_config.transform.use_modification_fcn = True
+    optim_config.transform.use_modification_fcn = False
 
-    if type == "sq":
+    if method == "sq":
         reference_config = calculation_config.model_copy(deep=True)
         reference_config.optimize = None
         reference_result = calculate_pdf(data_config, reference_config)
+        if min_range is None:
+            min_range = (0, reference_config.transform.q_max)
+
+    elif method == "gr" or method == "fr":
+        if min_range is None:
+            min_range = (0, optim_config.optimize.r_cutoff)
 
     def fcn(params):
         density = params["density"].value
+        bkg_scaling = params["bkg_scaling"].value
         optim_config.sample.density = density
+        data_config.bkg_scaling = bkg_scaling
         result = calculate_pdf(data_config, optim_config)
 
-        if type == "gr":
+        if method == "gr":
             r, gr = result.gr.limit(*min_range).data
-            residual = np.trapz(gr**2, r)
-        elif type == "sq":
+            residual = gr * (r[1] - r[0])
+        elif method == "fr":
+            atomic_density = optim_config.sample.atomic_density
+            r, fr = result.fr.limit(*min_range).data
+            residual = (fr + 4 * np.pi * r * atomic_density) * (r[1] - r[0])
+        elif method == "sq":
             q, sq = result.sq.limit(*min_range).data
             sq_ref = reference_result.sq.limit(*min_range).y
-            residual = np.trapz((sq - sq_ref) ** 2, q)
+            residual = (sq - sq_ref) * (q[1] - q[0])
+        else:
+            raise ValueError(
+                f"Invalid optimize density method: {method}, only 'gr', 'fr' and 'sq' are supported."
+            )
         return residual
 
-    if method == "nelder":
+    if optimization_method == "nelder":
         res = minimize(
             fcn,
             params,
             method="nelder",
             options={"maxfev": 500, "fatol": 0.0001, "xatol": 0.0001},
         )
-        return res.params["density"].value, res.residual[0]
-    elif method == "lsq":
+        return (
+            res.params["density"].value,
+            np.sum(res.residual**2),
+            res.params["bkg_scaling"].value,
+            np.sum(res.residual**2),
+        )
+    elif optimization_method == "lsq":
         res = minimize(
             fcn,
             params,
             method="least_squares",
         )
-        return res.params["density"].value, res.params["density"].stderr
+        return (
+            res.params["density"].value,
+            res.params["density"].stderr,
+            res.params["bkg_scaling"].value,
+            res.params["bkg_scaling"].stderr,
+        )
     else:
-        raise ValueError(f"Invalid optimization method: {method}")
+        raise ValueError(f"Invalid optimization method: {optimization_method}")