+++ changelog.md
@@ -1,3 +1,17 @@
+## 2.2.0 (2025/07/21)
+
+### New features
+
+- added xraylib scattering factor calculator, which can be used as an alternative to the Hajdu and Brown-Hubbell calculators - choose it by setting sf_source to 'xraylib' in the corresponding functions or in the configuration
+- optimize_density was reworked and extended. It now supports optimizing the density and background scaling at the same time. The "fr" method was added
+  and is the default method, since it the method described in Eggert et al. 2002 and other similar following publications.
+
+### Bugfixes
+
+- fixed an issue with the integral normalization method (normalize), which caused the normalization factor to be too high for most patterns
+- calculate_pdf now also applies the correct r_step and fourier_transform method to the optimize function
+- fixed a bug in the extrapolation functions (utility module), which was causing the a double point at min(x) of the extrapolated data (the one from the original data and one from the extrapolation)
+
 ## 2.1.0 (2025/07/12)
 
 ### New features